requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
boto3==1.18.53
chardet==3.0.4
certifi==2021.5.30
idna==3.2aniso8601==8.0.0
botocore==1.17.52
click==7.1.2
docutils==0.15.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
